Time: 0 ms (100%), Space: 40.2 MB (85.66%) - LeetHub

+++ 2649-count-total-number-of-colored-cells/2649-count-total-number-of-colored-cells.java
@@ -1,6 +1,6 @@
 class Solution {
     public long coloredCells(int n) {
-        long r = 1 + 4L*n*(n-1)/2;
+        long r = (long)(1 + 4L*n*(n-1)/2);
         return  r;
     }
 }
